backend-generation/pyproject.toml
[tool.poetry]
name = "terraform-backend-generation"
version = "0.1.0"
description = "Generates an Azure backend in terraform using j2 templates"
authors = ["gary <gary-stewart@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Jinja2 = "^3.1.2"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
